+++ changelog.md
@@ -18,6 +18,7 @@
   * Ability: Effect Spore.
 * The `/pokedex` now supports faster browsing! Press the "Return" button to view a list a Pokemon to scroll through!
 * You can now sort by Pokedex number in `/list`! Use `/list sortOrder: pokedexOrder`!
+* `/daily` now displays a timer until next reward if you've already claimed today!
 * **The Launch Celebration Event will be ending in 1 week! Make sure to pick up your perfect set of launch event legendaries before they go away!**
 
 ### 0.9.1 (6/1) - QoL Patch